Version 0.14 Alpha - Orc Room Rework!
This update is a complete rebuild of the Orc Room.  It is not currently fully functional, but full functionality will (probably) be coming in the next update.  Sorry if the updates seem a little slow, I have had minimal working time and need more time to get things done, but with the Thanksgiving break coming up, more features should see implementation in the near future.

(Aimed) Date of first official release: Unknown, maybe 2-3 months?  Lots of features to be added and not all variables have been accounted for yet.