Version 0.15 Alpha - Long Room Implemented!
The Orc Room is now almost complete and the Long Room has been reworked into the game (I tried to do it a while ago but it ruined the game for no reason :P, should work now).  To access it, go west from the Start Room.  It only has navigation tools so far, but it is fully functional and working.

If you find any issues, please contact me on Reddit (reddit.com/r/topdude155).  Thanks!