Version 0.16 Alpha - Bugfixes + New Rooms!
We have officially now hit 500 lines of code (in JavaScript alone)!  This is almost 30,000 characters of work, so I hope that my time hasn't been wasted and that someone will please enjoy my game.  We are well onto our way to a release by early next year, but no promises.  Once the mainframe for the game is complete, I intend to make some improvements to the overall gameplay and how real the game feels, because the game doesn't feel very in-depth at its current stage.  Thanks for playing!