Version 0.17 - UI Redesign!
Today I have added a new, more polished UI to the game.  I have also added a favicon (the little picture at the top when you are on the website).  Main thing to fix for the next update:  rewriting the whole area past the Army of Bones.  Expect them to not be functional in this update.