Version 0.18 Alpha - KEYWORDS!
Keywords have been implemented into this update!  They allow for better input recognition from the user, not needing exact inputs to trigger a result.  This should allow for the Chain Room to be better optimized for user experience, and will allow the code to be further condensed to improve page loading speeds.